:recycle: Refactor accounts management modal to be a reusable component
