+++ src/modules/Accounts/components/AccountsManagerModal/index.js
@@ -1,0 +1,27 @@
+import React from 'react';
+
+import { useTheme } from 'hooks/useTheme';
+import BottomModal from 'components/shared/BottomModal';
+import AccountItem from 'modules/Auth/components/AccountItem';
+import AccountsManager from '../AccountsManager';
+
+import getAccountsManagerModalStyles from './styles';
+
+export default function AccountsManagerModal({ show, setShow }) {
+  const { styles } = useTheme({ styles: getAccountsManagerModalStyles() });
+
+  return (
+    <BottomModal
+      show={show}
+      toggleShow={setShow}
+      style={{ container: styles.container }}
+    >
+      <AccountsManager
+        mode="modal"
+        onAccountSelect={() => setShow(false)}
+        item={AccountItem}
+        style={{ footer: styles.footer }}
+      />
+    </BottomModal>
+  );
+}
